refactor: 프로젝트 전체 리팩토링
