+++ app/src/main/java/com/yeolsimee/moneysaving/view/mypage/MyPageScreen.kt
@@ -33,7 +33,7 @@ import com.yeolsimee.moneysaving.utils.onClick
 @Composable
 fun MyPageScreen(
     alarmState: State<Boolean>,
-    onChangeAlarmState: () -> Unit = {},
+    onChangeAlarmState: (Boolean) -> Unit = {},
     onLogout: () -> Unit = {},
     onWithdraw: () -> Unit = {},
     openInternetBrowser: (String) -> Unit = {}
@@ -54,7 +54,7 @@ fun MyPageScreen(
             )
             Spacer(Modifier.height(17.dp))
             AlarmItem(alarmState) {
-                onChangeAlarmState()
+                onChangeAlarmState(it)
             }
             val agreementUrl = stringResource(R.string.agreement_url)
             val policyUrl = stringResource(R.string.policy_url)
@@ -91,12 +91,13 @@ fun MyPageScreen(
             onLeftClick = { onWithdraw() }
         )
     }
+
 }
 
 @Composable
 private fun AlarmItem(
     alarmState: State<Boolean>,
-    onChange: () -> Unit = {}
+    onChange: (Boolean) -> Unit = {}
 ) {
     Row(
         Modifier
@@ -114,7 +115,7 @@ private fun AlarmItem(
         Image(
             painter = painterResource(id = if (alarmState.value) R.drawable.toggle_on else R.drawable.toggle_off),
             contentDescription = "푸쉬 토글 버튼",
-            modifier = Modifier.onClick { onChange() }
+            modifier = Modifier.onClick { onChange(!alarmState.value) }
         )
     }
 }
@@ -161,19 +162,14 @@ fun MoveListItemPreview() {
 @Composable
 fun AlarmItemPreview() {
     Column {
-        AlarmItem(alarmState = remember {
+        val state = remember {
             mutableStateOf(false)
+        }
+        AlarmItem(alarmState = state, onChange = {
+            state.value = !state.value
         })
         AlarmItem(alarmState = remember {
             mutableStateOf(true)
         })
     }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun MyPageScreenPreview() {
-    MyPageScreen(
-        alarmState = remember { mutableStateOf(true) }
-    )
 }
