+++ app/src/main/java/com/yeolsimee/moneysaving/view/routine/GetRoutineViewModel.kt
@@ -36,6 +36,7 @@ class GetRoutineViewModel @Inject constructor(private val routineUseCase: Routin
 
     override fun showLoading() {
         intent {
+            reduce { RoutineResponse() }
             postSideEffect(ApiCallSideEffect.Loading)
         }
     }
