+++ app/src/main/java/com/yeolsimee/moneysaving/view/routine/RoutineScreen.kt
@@ -1,5 +1,3 @@
-@file:OptIn(ExperimentalMaterial3Api::class)
-
 package com.yeolsimee.moneysaving.view.routine
 
 import androidx.compose.foundation.background
@@ -11,12 +9,11 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
@@ -25,9 +22,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
 import com.yeolsimee.moneysaving.domain.calendar.AmPmTime
 import com.yeolsimee.moneysaving.domain.calendar.getCurrentTime
-import com.yeolsimee.moneysaving.domain.entity.category.TextItem
 import com.yeolsimee.moneysaving.domain.entity.routine.RoutineRequest
 import com.yeolsimee.moneysaving.domain.entity.routine.RoutineResponse
 import com.yeolsimee.moneysaving.ui.appbar.BottomButtonAppBar
@@ -41,31 +38,32 @@ import com.yeolsimee.moneysaving.utils.getTwoDigits
 import com.yeolsimee.moneysaving.utils.getWeekTypeCheckList
 import com.yeolsimee.moneysaving.utils.getWeekTypes
 import com.yeolsimee.moneysaving.view.category.CategoryGridView
+import com.yeolsimee.moneysaving.view.category.CategoryViewModel
 
-@ExperimentalMaterial3Api
-@ExperimentalLayoutApi
 @Composable
 fun RoutineScreen(
     routine: RoutineResponse = RoutineResponse(),
-    routineType: RoutineModifyOption? = RoutineModifyOption.Update,
-    categoryList: MutableList<TextItem> = mutableListOf(),
+    routineType: RoutineType? = RoutineType.Update,
+    categoryViewModel: CategoryViewModel = hiltViewModel(),
     selectedCategoryId: MutableState<String> = remember { mutableStateOf("") },
-    notificationCheckDialogState: MutableState<Boolean> = remember { mutableStateOf(false) },
     closeCallback: () -> Unit = {},
-    onCompleteCallback: (req: RoutineRequest, hasWeekTypes: Boolean) -> Unit = { _, _ -> },
-    toggleRoutineAlarm: (MutableState<Boolean>) -> Unit = {},
-    onCategoryAdded: (String) -> Unit = {},
-    onCheckNotificationSetting: () -> Unit = {}
+    onCompleteCallback: (req: RoutineRequest, hasWeekTypes: Boolean, routine: RoutineResponse) -> Unit = { _, _, _ -> },
+    toggleRoutineAlarm: (MutableState<Boolean>, MutableState<Boolean>) -> Unit = { _, _ -> },
+    onCheckNotificationSetting: () -> Unit = {},
 ) {
+    val notificationCheckDialogState = remember { mutableStateOf(false) }
+
     val buttonState = remember { mutableStateOf(false) }
     val timePickerDialogState = remember { mutableStateOf(false) }
+
+    val categoryList = categoryViewModel.container.stateFlow.collectAsState().value
 
     RoumoTheme(navigationBarColor = if (buttonState.value) Color.Black else Gray99) {
         val focusRequester by remember { mutableStateOf(FocusRequester()) }
         val focusManager = LocalFocusManager.current
         val scrollState = rememberScrollState()
 
-        val routineName = remember { mutableStateOf(routine.routineName) }
+        val nameState = remember { mutableStateOf(routine.routineName) }
 
         if (categoryList.isNotEmpty() && selectedCategoryId.value.isEmpty()) {
             selectedCategoryId.value = categoryList.first().id
@@ -79,12 +77,12 @@ fun RoutineScreen(
         val now = getCurrentTime()
 
         val hourState: MutableState<Int> = remember {
-            mutableIntStateOf(
+            mutableStateOf(
                 if (alarmState.value) routine.alarmTime.substring(0, 2).toInt() else now.first
             )
         }
         val minuteState: MutableState<Int> = remember {
-            mutableIntStateOf(
+            mutableStateOf(
                 if (alarmState.value) routine.alarmTime.substring(2, 4).toInt() else now.second
             )
         }
@@ -96,20 +94,25 @@ fun RoutineScreen(
                 TopBackButtonTitleAppBar(routineType?.title) { closeCallback() }
             },
             bottomBar = {
-                buttonState.value = canSaveRoutine(routineName.value, selectedCategoryId)
+                buttonState.value = canSaveRoutine(nameState.value, selectedCategoryId.value)
 
                 BottomButtonAppBar(routineType?.title, buttonState.value) {
                     val weekTypes = getWeekTypes(repeatSelectList)
                     onCompleteCallback(
                         RoutineRequest(
                             alarmStatus = if (alarmState.value) "ON" else "OFF",
-                            alarmTime = getAlarmTime(alarmState, hourState, minuteState),
-                            routineName = routineName.value,
+                            alarmTime = getAlarmTime(
+                                alarmState.value,
+                                hourState.value,
+                                minuteState.value
+                            ),
+                            routineName = nameState.value,
                             categoryId = selectedCategoryId.value,
                             routineTimeZone = selectedRoutineTimeZoneId.value,
                             weekTypes = weekTypes,
                         ),
-                        weekTypes.isNotEmpty()
+                        weekTypes.isNotEmpty(),
+                        routine
                     )
                 }
             },
@@ -123,17 +126,17 @@ fun RoutineScreen(
                     .addFocusCleaner(focusManager)
             ) {
                 Column(Modifier.verticalScroll(scrollState)) {
-                    InputRoutineName(routineName, focusRequester)
+                    InputRoutineName(nameState, focusRequester)
                     Spacer(Modifier.height(20.dp))
                     CategoryGridView(
                         categoryList,
-                        selectedId = selectedCategoryId,
+                        selectedId = selectedCategoryId.value,
                         addCategoryState = addCategoryState,
                         selectCallback = { id ->
                             selectedCategoryId.value = id
                         },
                         addCallback = { categoryName ->
-                            onCategoryAdded(categoryName)
+                            categoryViewModel.addCategory(categoryName)
                         }
                     )
                     Spacer(Modifier.height(20.dp))
@@ -142,7 +145,7 @@ fun RoutineScreen(
                     }
                     Spacer(Modifier.height(20.dp))
                     SelectRoutineTimeZone(
-                        selectedId = selectedRoutineTimeZoneId
+                        selectedId = selectedRoutineTimeZoneId.value
                     ) { id ->
                         selectedRoutineTimeZoneId.value = id
                     }
@@ -152,7 +155,8 @@ fun RoutineScreen(
                         hourState,
                         minuteState,
                         toggleRoutineAlarm,
-                        timePickerDialogState
+                        timePickerDialogState,
+                        notificationCheckDialogState
                     )
                 }
             }
@@ -193,37 +197,28 @@ private fun getRoutineRepeatList(initialData: RoutineResponse?) =
 @Composable
 private fun canSaveRoutine(
     routineName: String,
-    selectedCategoryId: MutableState<String>
-) = routineName.isNotEmpty() && selectedCategoryId.value.isNotEmpty()
+    categoryId: String
+) = routineName.isNotEmpty() && categoryId.isNotEmpty()
 
 private fun getAlarmTime(
-    alarmState: MutableState<Boolean>,
-    hourState: MutableState<Int>,
-    minuteState: MutableState<Int>
+    alarmState: Boolean,
+    hourState: Int,
+    minuteState: Int
 ): String {
-    return if (alarmState.value) {
-        val hour = hourState.value.getTwoDigits()
-        val minute = minuteState.value.getTwoDigits()
+    return if (alarmState) {
+        val hour = hourState.getTwoDigits()
+        val minute = minuteState.getTwoDigits()
         "${hour}${minute}"
     } else {
         ""
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
 @ExperimentalLayoutApi
 @Preview(showBackground = true)
 @Composable
 fun RoutineAddScreenPreview() {
     RoutineScreen(
-        routineType = RoutineModifyOption.Add,
-        categoryList = remember {
-            mutableListOf(
-                TextItem("1", "💰아껴쓰기"),
-                TextItem("2", "주린이 성장일기"),
-                TextItem("3", "임티는 사용자 자유"),
-                TextItem("4", "열네글자까지들어가요일이삼사")
-            )
-        },
+        routineType = RoutineType.Add,
     )
 }
