+++ app/src/main/java/com/yeolsimee/moneysaving/view/routine/RoutineType.kt
@@ -3,7 +3,7 @@ package com.yeolsimee.moneysaving.view.routine
 import androidx.annotation.Keep
 
 @Keep
-enum class RoutineModifyOption(val title: String) {
+enum class RoutineType(val title: String) {
     Add("루틴 추가하기"),
     Update("루틴 수정하기")
 }
