+++ data/build.gradle.kts
@@ -51,7 +51,7 @@ android {
 dependencies {
     implementation(project(":domain"))
 
-    val hiltVersion = "2.46.1"
+    val hiltVersion = "2.47"
     implementation("androidx.core:core-ktx:1.10.1")
 
     // Firebase Auth
@@ -67,7 +67,7 @@ dependencies {
     kapt("com.google.dagger:hilt-android-compiler:$hiltVersion")
 
     // Coroutine
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.1")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3")
 
     // room DB
     val roomVersion = "2.5.2"
@@ -82,7 +82,7 @@ dependencies {
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
-    testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.7.1")
+    testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.7.3")
     testImplementation("androidx.arch.core:core-testing:2.2.0")
 
     // kotest
